refactor(flask): refactor all (wip)
