+++ src/emoji/routes/api/history.py
@@ -1,0 +1,1 @@
+# -*- encoding: utf-8 -*-
